CNC internal state revision
 - Internal states revision (8bit to 16bit)
 - added 3 new extended Grbl states (optional) (Dweel, Probe and Locked)
 - Added skeleton for Security Door Parking

+++ uCNC/src/core/interpolator.c
@@ -838,7 +838,7 @@ MCU_CALLBACK void itp_stop(void)
 	// safer to make the stoping condition block
 	ATOMIC_CODEBLOCK
 	{
-		uint8_t state = cnc_get_exec_state(EXEC_ALLACTIVE);
+		uint16_t state = cnc_get_exec_state(EXEC_ALLACTIVE);
 
 		// any stop command while running triggers an HALT alarm
 		if (state & EXEC_RUN)
